Implement activity to show open source libs used
