+++ app/build.gradle
@@ -6,6 +6,7 @@ apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
+apply plugin: 'com.google.gms.oss.licenses.plugin'
 
 android {
     compileSdkVersion rootProject.compileSdkVersion
@@ -61,6 +62,7 @@ dependencies {
     implementation "com.android.support:support-v4:$supportVersion"
 
     implementation "com.google.firebase:firebase-messaging:$firebaseVersion"
+    implementation "com.google.android.gms:play-services-oss-licenses:$firebaseVersion"
 
     implementation "androidx.core:core-ktx:$ktxVersion"
