+++ app/src/main/java/com/oskhoj/swingplanner/ui/component/OnTouchAnimation.kt
@@ -9,8 +9,8 @@ import com.oskhoj.swingplanner.util.getLong
 class OnTouchAnimation : View.OnTouchListener {
     override fun onTouch(view: View, event: MotionEvent): Boolean {
         when (event.action) {
-            MotionEvent.ACTION_DOWN -> view.animateToSize(0.95f, view.context.getLong(R.integer.anim_duration_medium))
-            MotionEvent.ACTION_CANCEL -> view.animateToSize(1f, view.context.getLong(R.integer.anim_duration_medium))
+            MotionEvent.ACTION_DOWN -> view.animateToSize(0.95f, view.getLong(R.integer.anim_duration_medium))
+            MotionEvent.ACTION_CANCEL -> view.animateToSize(1f, view.getLong(R.integer.anim_duration_medium))
         }
         return false
     }
