+++ app/src/main/res/raw/third_party_license_metadata
@@ -1,0 +1,116 @@
+0:47 firebase-core
+48:47 firebase-analytics
+96:47 firebase-analytics-impl
+144:47 firebase-iid
+192:47 firebase-common
+240:47 play-services-tasks
+288:47 play-services-basement
+336:2500 zlib
+2837:1602 JSR 305
+4440:11358 safeparcel
+15799:3182 PCRE
+18982:680 STL
+19663:731 UTF
+20395:1558 RE2
+21954:1481 darts_clone
+23436:1732 Protobuf Nano
+25169:19442 ICU4C
+44612:11342 flatbuffers
+55955:11358 CCTZ
+67314:243 tz database
+67558:46 kotlin-android-extensions-runtime
+72334:46 kotlin-stdlib-jre7
+67652:34 kodein-android
+68012:46 customtabs
+67734:42 appintro
+67777:46 taptargetview
+67824:46 rxbinding-design
+67871:46 design
+67918:46 appcompat-v7
+67965:46 cardview-v7
+68059:46 recyclerview-v7
+68106:46 core-ktx
+68153:46 conductor
+68200:46 rxbinding
+68247:92 glide
+68340:47 firebase-messaging
+68388:47 play-services-oss-licenses
+68436:46 support-v4
+68483:46 support-fragment
+68530:46 support-core-utils
+68577:46 animated-vector-drawable
+68624:46 support-core-ui
+68671:92 gifdecoder
+68764:46 support-vector-drawable
+68811:46 transition
+68858:46 support-media-compat
+68905:46 support-compat
+68952:46 support-annotations
+68999:31 store-kotlin3
+69031:31 middleware3
+69063:31 filesystem3
+69095:31 store3
+69127:46 rxandroid
+69915:46 rxjava
+69221:46 kotpref
+69268:46 timber
+69315:47 installreferrer
+69363:45 anko-sdk21
+69409:45 anko-appcompat-v7
+69455:45 anko-sdk21-listeners
+69501:45 anko-design
+69547:34 kodein
+69582:34 kodein-conf
+69617:34 kodein-core
+69652:45 anko-support-v4
+69698:45 anko-appcompat-v7-commons
+69744:45 anko-support-v4-commons
+69790:45 anko-commons
+72381:46 kotlin-stdlib
+69883:31 cache3
+69962:92 disklrucache
+72118:46 annotations
+70148:49 reactive-streams
+70245:46 runtime
+71831:46 common
+70386:46 kotlin-annotation-processing-gradle
+70433:46 kotlin-compiler-embeddable
+71232:46 kotlin-reflect
+70527:46 kotlin-script-runtime
+70574:46 lint-gradle
+70621:46 builder
+70668:46 lint
+70715:46 lint-checks
+70762:46 lint-api
+70809:46 manifest-merger
+70856:46 sdk-common
+70903:46 lint-gradle-api
+70950:46 gradle-api
+70997:46 builder-model
+71044:46 lint-kotlin
+71091:46 intellij-core
+71138:46 uast
+71185:46 groovy-all
+71279:46 sdklib
+71326:46 repository
+71373:46 kotlin-stdlib-jre8
+71420:46 builder-test-api
+71467:46 ddmlib
+71514:40 bcpkix-jdk15on
+71555:40 bcprov-jdk15on
+71596:46 tracker
+71643:46 shared
+71690:46 protos
+71737:46 layoutlib-api
+71784:46 dvlib
+71878:46 apksig
+71925:46 javawriter
+71972:47 fastutil
+72020:50 jopt-simple
+72071:46 json-simple
+72165:121 kxml2
+72287:46 kotlin-compiler
+72428:46 jsr305
+72475:46 j2objc-annotations
+72522:55 stax-ex
+72578:102 jaxb-api
