+++ build.gradle
@@ -43,6 +43,7 @@ buildscript {
         classpath 'com.android.tools.build:gradle:3.1.0-alpha09'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
         classpath 'com.google.gms:google-services:3.1.1'
+        classpath "com.google.gms:oss-licenses:0.9.1"
     }
 }
